refactor: decompose inventory into backpack/hotbar
